Runge Kutta

RungeKutta.Tableau

Base.show

RungeKutta.TableauERK416

RungeKutta.TableauERK438

RungeKutta.TableauExplicitEuler

RungeKutta.TableauExplicitMidpoint

RungeKutta.TableauGauss

RungeKutta.TableauHeun

RungeKutta.TableauImplicitEuler

RungeKutta.TableauImplicitMidpoint

RungeKutta.TableauKutta

RungeKutta.TableauLobattoIIIA

RungeKutta.TableauLobattoIIIB

RungeKutta.TableauLobattoIIIC

RungeKutta.TableauLobattoIIIC̄

RungeKutta.TableauLobattoIIID

RungeKutta.TableauLobattoIIIE

RungeKutta.TableauLobattoIIIF

RungeKutta.TableauLobattoIIIG

RungeKutta.TableauRadauIIA

RungeKutta.TableauRunge

RungeKutta.TableauSRK3

RungeKutta.get_gauss_coefficients

RungeKutta.get_gauss_nodes

RungeKutta.get_gauss_weights

RungeKutta.get_lobatto_coefficients_a

RungeKutta.get_lobatto_coefficients_b

RungeKutta.get_lobatto_coefficients_c

RungeKutta.get_lobatto_coefficients_c̄

RungeKutta.get_lobatto_nodes

RungeKutta.get_lobatto_weights

RungeKutta.get_radau_coefficients

RungeKutta.get_radau_nodes

RungeKutta.get_radau_weights

RungeKutta.solve_simplifying_assumption_b

RungeKutta.solve_simplifying_assumption_c

RungeKutta.solve_simplifying_assumption_d

RungeKutta.Tableau — Type

Holds the tableau of a Runge-Kutta method.

Base.show — Method

Print Runge-Kutta coefficients.

RungeKutta.TableauERK416 — Function

Tableau for explicit Runge-Kutta method of order four (1/6 rule)

RungeKutta.TableauERK438 — Function

Tableau for explicit Runge-Kutta method of order four (3/8 rule)

RungeKutta.TableauExplicitEuler — Function

Tableau for explicit Euler method

RungeKutta.TableauExplicitMidpoint — Function

Tableau for explicit midpoint method

RungeKutta.TableauGauss — Function

Gauss tableau with s stages

RungeKutta.TableauHeun — Function

Tableau for Heun's method

RungeKutta.TableauImplicitEuler — Function

Implicit Euler

RungeKutta.TableauImplicitMidpoint — Function

Implicit Midpoint

RungeKutta.TableauKutta — Function

Tableau for Kutta's method of order three

RungeKutta.TableauLobattoIIIA — Function

Lobatto IIIA tableau with s stages

RungeKutta.TableauLobattoIIIB — Function

Lobatto IIIB tableau with s stages

RungeKutta.TableauLobattoIIIC — Function

Lobatto IIIC tableau with s stages

RungeKutta.TableauLobattoIIIC̄ — Function

Lobatto IIIC̄ tableau with s stages

RungeKutta.TableauLobattoIIID — Function

Lobatto IIID tableau with s stages

RungeKutta.TableauLobattoIIIE — Function

Lobatto IIIE tableau with s stages

RungeKutta.TableauLobattoIIIF — Function

Lobatto IIIF tableau with s stages

RungeKutta.TableauLobattoIIIG — Function

Lobatto IIIG tableau with s stages

RungeKutta.TableauRadauIIA — Function

Radau IIA tableau with s stages

RungeKutta.TableauRunge — Function

Tableau for Runge's method

RungeKutta.TableauSRK3 — Function

Symmetric Runge-Kutta tableau with three stages.

RungeKutta.get_gauss_coefficients — Method

The Gauss coefficients are implicitly given by the so-called simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_gauss_nodes — Method

The Gauss nodes are given by the roots of the shifted Legendre polynomial $P_s (2x-1)$ with $s$ the number of stages.

RungeKutta.get_gauss_weights — Method

The Gauss weights are given by the following integrals

\[b_i = \bigg( \frac{dP}{dx} (c_i) \bigg)^{-2} \int \limits_0^1 \bigg( \frac{P(x)}{x - c_i} \bigg)^2 dx ,\]

where $P(x)$ denotes the shifted Legendre polynomial $P(x) = P_s (2x-1)$ with $s$ the number of stages.

RungeKutta.get_lobatto_coefficients_a — Method

The Lobatto IIIA coefficients are implicitly given by the so-called simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_lobatto_coefficients_b — Method

The Lobatto IIIB coefficients are implicitly given by the so-called simplifying assumption $D(s)$:

\[\sum \limits_{i=1}^{s} b_i c_{i}^{k-1} a_{ij} = \frac{b_j}{k} ( 1 - c_j^k) \qquad j = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_lobatto_coefficients_c — Method

The Lobatto IIIC coefficients are determined by setting $a_{i,1} = b_1$ and solving the so-called simplifying assumption $C(s-1)$, given by

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s-1 ,\]

for $a_{i,j}$ with $i = 1, ..., s$ and $j = 2, ..., s$.

RungeKutta.get_lobatto_coefficients_c̄ — Method

The Lobatto IIIC̄ coefficients are determined by setting $a_{i,s} = 0$ and solving the so-called simplifying assumption $C(s-1)$, given by

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s-1 ,\]

for $a_{i,j}$ with $i = 1, ..., s$ and $j = 1, ..., s-1$.

RungeKutta.get_lobatto_nodes — Method

The s-stage Lobatto nodes are defined as the roots of the following polynomial of degree $s$:

\[\frac{d^{s-2}}{dx^{s-2}} \big( (x - x^2)^{s-1} \big) .\]

RungeKutta.get_lobatto_weights — Method

The Lobatto weights can be explicitly computed by the formula

\[b_j = \frac{1}{s (s-1) P_{s-1}(2 c_j - 1)^2} \qquad j = 1 , \, ... , \, s ,\]

where $P_k$ is the $k$th Legendre polynomial, given by

\[P_k (x) = \frac{1}{k! 2^k} \big( \frac{d^k}{dx^k} (x^2 - 1)^k \big) .\]

RungeKutta.get_radau_coefficients — Method

The Radau IIA coefficients are implicitly given by the so-called simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_radau_nodes — Method

The s-stage Radau nodes are defined as the roots of the following polynomial of degree $s$:

\[\frac{d^{s-2}}{dx^{s-2}} \big( (x - x^2)^{s-1} \big) .\]

RungeKutta.get_radau_weights — Method

The Radau weights are implicitly given by the so-called simplifying assumption $B(s)$:

\[\sum \limits_{j=1}^{s} b_{j} c_{j}^{k-1} = \frac{1}{k} \qquad k = 1 , \, ... , \, s .\]

RungeKutta.solve_simplifying_assumption_b — Method

Compute the weights by solving the simplifying assumption $B(s)$:

\[\sum \limits_{j=1}^{s} b_{j} c_{j}^{k-1} = \frac{1}{k} \qquad k = 1 , \, ... , \, s .\]

RungeKutta.solve_simplifying_assumption_c — Method

Compute the coefficients by solving the simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.solve_simplifying_assumption_d — Method

Compute the coefficients by solving the simplifying assumption $D(s)$:

\[\sum \limits_{i=1}^{s} b_i c_{i}^{k-1} a_{ij} = \frac{b_j}{k} ( 1 - c_j^k) \qquad j = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]