Runge Kutta Library Functions

RungeKutta.Tableau — Type

Holds the tableau of a Runge-Kutta method.

Parameters:

T: datatype of coefficient arrays

Fields:

name: symbolic name of the tableau

o: order of the method

s: number of stages

a: coefficients $a_{ij}$ with $ 1 \le i,j \le s$

b: weights $b_{i}$ with $ 1 \le i \le s$

c: nodes $c_{i}$ with $ 1 \le i \le s$

Base.show — Method

Markdown-print Runge-Kutta tableau.

Base.show — Method

Pretty-print Runge-Kutta tableau.

RungeKutta.TableauBackwardEuler — Function

Alias for TableauImplicitEuler

RungeKutta.TableauCrankNicolson — Method

Tableau of Crank-Nicolson two-stage, 2nd order method

Reference: J. Crank and P. Nicolson. A practical method for numerical evaluation of solutions of partial differential equations of the heat-conduction type. Mathematical Proceedings of the Cambridge Philosophical Society, Volume 43, Issue 1, Pages 50-67, 1947. doi: 10.1017/S0305004100023197

RungeKutta.TableauCrouzeix — Method

Tableau of Crouzeix's two-stage, 3rd order method

Reference: M.Crouzeix. Sur L'approximation des équations différentielles opérationelles linéaires par des méthodes de Runge-Kutta. Thesis. Université de Paris, 1975.

RungeKutta.TableauExplicitEuler — Method

Tableau of one-stage, 1st order explicit (forward) Euler method

Reference: Leonhard Euler Institutiones calculi differentialis cum eius vsu in analysi finitorum ac doctrina serierum. Imp. Acad. Imper. Scient. Petropolitanae, Opera Omnia, Vol.X, [I.6], 1755.

RungeKutta.TableauExplicitMidpoint — Method

Tableau of explicit two-stage, 2nd order midpoint method

RungeKutta.TableauForwardEuler — Function

Alias for TableauExplicitEuler

RungeKutta.TableauGauss — Method

Gauss tableau with s stages

RungeKutta.TableauHeun2 — Method

Tableau of Heun's two-stage, 2nd order method

Reference: Karl Heun. Neue Methoden zur approximativen Integration der Differentialgleichungen einer unabhängigen Veränderlichen. Zeitschrift für Mathematik und Physik, Volume 45, Pages 23-38, 1900. Algorithm II.

RungeKutta.TableauHeun3 — Method

Tableau of Heun's three-stage, 3rd order method

Reference: Karl Heun. Neue Methoden zur approximativen Integration der Differentialgleichungen einer unabhängigen Veränderlichen. Zeitschrift für Mathematik und Physik, Volume 45, Pages 23-38, 1900. Algorithm VI.

RungeKutta.TableauImplicitEuler — Method

Tableau of one-stage, 1st order implicit (backward) Euler method

RungeKutta.TableauImplicitMidpoint — Method

Tableau of two-stage, 2nd order implicit midpoint method

RungeKutta.TableauKraaijevangerSpijker — Method

Tableau of Kraaijevanger and Spijker's two-stage, 2nd order method

Reference: J. F. B. M. Kraaijevanger and M. N. Spijker. Algebraic stability and error propagation in Runge-Kutta methods. Applied Numerical Mathematics, Volume 5, Issues 1-2, Pages 71-87, 1989. doi: 10.1016/0168-9274(89)90025-1

RungeKutta.TableauKutta — Method

Tableau of Kutta's three-stage, 3rd order method

Reference: Wilhelm Kutta Beitrag zur Näherungsweisen Integration totaler Differentialgleichungen Zeitschrift für Mathematik und Physik, Volume 46, Pages 435–453, 1901. Page 440

RungeKutta.TableauKutta3 — Function

Alias for TableauKutta

RungeKutta.TableauLobattoIIIA — Method

Lobatto IIIA tableau with s stages

RungeKutta.TableauLobattoIIIB — Method

Lobatto IIIB tableau with s stages

RungeKutta.TableauLobattoIIIC — Method

Lobatto IIIC tableau with s stages

RungeKutta.TableauLobattoIIIC̄ — Method

Lobatto IIIC̄ tableau with s stages

RungeKutta.TableauLobattoIIID — Method

Lobatto IIID tableau with s stages

RungeKutta.TableauLobattoIIIE — Method

Lobatto IIIE tableau with s stages

RungeKutta.TableauLobattoIIIF — Method

Lobatto IIIF tableau with s stages

RungeKutta.TableauLobattoIIIG — Method

Lobatto IIIG tableau with s stages

RungeKutta.TableauQinZhang — Method

Tableau of Qin and Zhang's symplectic two-stage, 2nd order method

Reference: M.-Z. Qin and M.-Q. Zhang. Symplectic Runge-Kutta algorithms for Hamilton systems. Journal of Computational Mathematics, Supplementary Issue, Pages 205-215, 1992.

RungeKutta.TableauRK4 — Function

Alias for TableauRK416

RungeKutta.TableauRK416 — Method

Tableau of explicit Runge-Kutta method of order four (1/6 rule)

Reference: Wilhelm Kutta Beitrag zur Näherungsweisen Integration totaler Differentialgleichungen Zeitschrift für Mathematik und Physik, Volume 46, Pages 435–453, 1901. Page 443

RungeKutta.TableauRK438 — Method

Tableau of explicit Runge-Kutta method of order four (3/8 rule)

Reference: Wilhelm Kutta Beitrag zur Näherungsweisen Integration totaler Differentialgleichungen Zeitschrift für Mathematik und Physik, Volume 46, Pages 435–453, 1901. Page 441

RungeKutta.TableauRadauIA — Method

Radau IA tableau with s stages

RungeKutta.TableauRadauIIA — Method

Radau IIA tableau with s stages

RungeKutta.TableauRalston2 — Method

Tableau of Ralston's two-stage, 2nd order method

Reference: Anthony Ralston. Runge-Kutta Methods with Minimum Error Bounds. Mathematics of Computation, Volume 16, Pages 431-437, 1962. doi: 10.1090/S0025-5718-1962-0150954-0 Equation (3.5)

RungeKutta.TableauRalston3 — Method

Tableau of Ralston's three-stage, 3rd order method

Reference: Anthony Ralston. Runge-Kutta Methods with Minimum Error Bounds. Mathematics of Computation, Volume 16, Pages 431-437, 1962. doi: 10.1090/S0025-5718-1962-0150954-0 Equation (4.10)

RungeKutta.TableauRunge — Method

Tableau of Runge's two-stage, 2nd order method

Referene: Carl Runge Über die numerische Auflösung von Differentialgleichungen. Mathematische Annalen, Volume 46, Pages 167-178, 1895. doi: 10.1007/BF01446807 Equation (3)

RungeKutta.TableauRunge2 — Function

Alias for TableauRunge

RungeKutta.TableauSRK3 — Method

Tableau of symmetric and symplectic three-stage, 4th order Runge-Kutta method

Reference: Shan Zhao and Guo-Wei Wei. A unified discontinuous Galerkin framework for time integration. Mathematical Methods in the Applied Sciences, Volume 37, Issue 7, Pages 1042-1071, 2014. doi: 10.1002/mma.2863.

RungeKutta.TableauSSPRK2 — Function

Tableau of 2rd order Strong Stability Preserving method with two stages and CFL ≤ 1

Reference: Chi-Wang Shu, Stanley Osher. Efficient implementation of essentially non-oscillatory shock-capturing schemes. Journal of Computational Physics, Volume 77, Issue 2, Pages 439-471, 1988. Equation (2.16)

Alias for TableauHeun2

RungeKutta.TableauSSPRK3 — Method

Tableau of 3rd order Strong Stability Preserving method with three stages and CFL ≤ 1

Reference: Chi-Wang Shu, Stanley Osher. Efficient implementation of essentially non-oscillatory shock-capturing schemes. Journal of Computational Physics, Volume 77, Issue 2, Pages 439-471, 1988. Equation (2.18)

RungeKutta._legendre — Method

Legendre polynomial P_s(x) of degree s defined on the interval [-1..+1].

RungeKutta._shifted_legendre — Function

Legendre polynomial of degree s shifted to the interval [0..1], i.e., P_s(2x-1).

RungeKutta.get_gauss_coefficients — Method

The Gauss coefficients are implicitly given by the so-called simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_gauss_nodes — Method

The Gauss nodes are given by the roots of the shifted Legendre polynomial $P_s (2x-1)$ with $s$ the number of stages.

RungeKutta.get_gauss_weights — Method

The Gauss weights are given by the following integrals

\[b_i = \bigg( \frac{dP}{dx} (c_i) \bigg)^{-2} \int \limits_0^1 \bigg( \frac{P(x)}{x - c_i} \bigg)^2 dx ,\]

where $P(x)$ denotes the shifted Legendre polynomial $P(x) = P_s (2x-1)$ with $s$ the number of stages.

RungeKutta.get_lobatto_a_coefficients — Method

The Lobatto IIIA coefficients are implicitly given by the so-called simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_lobatto_b_coefficients — Method

The Lobatto IIIB coefficients are implicitly given by the so-called simplifying assumption $D(s)$:

\[\sum \limits_{i=1}^{s} b_i c_{i}^{k-1} a_{ij} = \frac{b_j}{k} ( 1 - c_j^k) \qquad j = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_lobatto_c_coefficients — Method

The Lobatto IIIC coefficients are determined by setting $a_{i,1} = b_1$ and solving the so-called simplifying assumption $C(s-1)$, given by

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s-1 ,\]

for $a_{i,j}$ with $i = 1, ..., s$ and $j = 2, ..., s$.

RungeKutta.get_lobatto_c̄_coefficients — Method

The Lobatto IIIC̄ coefficients are determined by setting $a_{i,s} = 0$ and solving the so-called simplifying assumption $C(s-1)$, given by

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s-1 ,\]

for $a_{i,j}$ with $i = 1, ..., s$ and $j = 1, ..., s-1$.

RungeKutta.get_lobatto_nodes — Method

The s-stage Lobatto nodes are defined as the roots of the following polynomial of degree $s$:

\[\frac{d^{s-2}}{dx^{s-2}} \big( (x - x^2)^{s-1} \big) .\]

RungeKutta.get_lobatto_weights — Method

The Lobatto weights can be explicitly computed by the formula

\[b_j = \frac{1}{s (s-1) P_{s-1}(2 c_j - 1)^2} \qquad j = 1 , \, ... , \, s ,\]

where $P_k$ is the $k$th Legendre polynomial, given by

\[P_k (x) = \frac{1}{k! 2^k} \big( \frac{d^k}{dx^k} (x^2 - 1)^k \big) .\]

RungeKutta.get_radau_1_coefficients — Method

The Radau IA coefficients are implicitly given by the so-called simplifying assumption $D(s)$:

\[\sum \limits_{i=1}^{s} b_i c_{i}^{k-1} a_{ij} = \frac{b_j}{k} ( 1 - c_j^k) \qquad j = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_radau_1_nodes — Method

The s-stage Radau IA nodes are defined as the roots of the following polynomial of degree $s$:

\[\frac{d^{s-1}}{dx^{s-1}} \big( x^s (x - 1)^{s-1} \big) .\]

RungeKutta.get_radau_1_weights — Method

The Radau IA weights are implicitly given by the so-called simplifying assumption $B(s)$:

\[\sum \limits_{j=1}^{s} b_{j} c_{j}^{k-1} = \frac{1}{k} \qquad k = 1 , \, ... , \, s .\]

RungeKutta.get_radau_2_coefficients — Method

The Radau IIA coefficients are implicitly given by the so-called simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_radau_2_nodes — Method

The s-stage Radau IIA nodes are defined as the roots of the following polynomial of degree $s$:

\[\frac{d^{s-1}}{dx^{s-1}} \big( x^{s-1} (x - 1)^s \big) .\]

RungeKutta.get_radau_2_weights — Method

The Radau IIA weights are implicitly given by the so-called simplifying assumption $B(s)$:

\[\sum \limits_{j=1}^{s} b_{j} c_{j}^{k-1} = \frac{1}{k} \qquad k = 1 , \, ... , \, s .\]

RungeKutta.solve_simplifying_assumption_b — Method

Compute the weights by solving the simplifying assumption $B(s)$:

\[\sum \limits_{j=1}^{s} b_{j} c_{j}^{k-1} = \frac{1}{k} \qquad k = 1 , \, ... , \, s .\]

RungeKutta.solve_simplifying_assumption_c — Method

Compute the coefficients by solving the simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.solve_simplifying_assumption_d — Method

Compute the coefficients by solving the simplifying assumption $D(s)$:

\[\sum \limits_{i=1}^{s} b_i c_{i}^{k-1} a_{ij} = \frac{b_j}{k} ( 1 - c_j^k) \qquad j = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]