Runge Kutta Library Functions

RungeKutta.Tableau — Type

Holds the tableau of a Runge-Kutta method.

Base.show — Method

Pretty-print Runge-Kutta tableau.

RungeKutta.TableauCrankNicolson — Function

Tableau of Crank-Nicolson two-stage, 2nd order method

RungeKutta.TableauCrouzeix — Function

Tableau of Crouzeix's two-stage, 3rd order method

RungeKutta.TableauExplicitEuler — Function

Tableau of one-stage, 1st order explicit (forward) Euler method

RungeKutta.TableauExplicitMidpoint — Function

Tableau of explicit two-stage, 2nd order midpoint method

RungeKutta.TableauGauss — Function

Gauss tableau with s stages

RungeKutta.TableauHeun2 — Function

Tableau of Heun's two-stage, 2nd order method

RungeKutta.TableauHeun3 — Function

Tableau of Heun's three-stage, 3rd order method

RungeKutta.TableauImplicitEuler — Function

Tableau of one-stage, 1st order implicit (backward) Euler method

RungeKutta.TableauImplicitMidpoint — Function

Tableau of two-stage, 2nd order implicit midpoint method

RungeKutta.TableauKraaijevangerSpijker — Function

Tableau of Kraaijevanger and Spijker's two-stage, 2nd order method

RungeKutta.TableauKutta — Function

Tableau of Kutta's three-stage, 3rd order method

RungeKutta.TableauLobattoIIIA — Function

Lobatto IIIA tableau with s stages

RungeKutta.TableauLobattoIIIB — Function

Lobatto IIIB tableau with s stages

RungeKutta.TableauLobattoIIIC — Function

Lobatto IIIC tableau with s stages

RungeKutta.TableauLobattoIIIC̄ — Function

Lobatto IIIC̄ tableau with s stages

RungeKutta.TableauLobattoIIID — Function

Lobatto IIID tableau with s stages

RungeKutta.TableauLobattoIIIE — Function

Lobatto IIIE tableau with s stages

RungeKutta.TableauLobattoIIIF — Function

Lobatto IIIF tableau with s stages

RungeKutta.TableauLobattoIIIG — Function

Lobatto IIIG tableau with s stages

RungeKutta.TableauQinZhang — Function

Tableau of Qin and Zhang's symplectic two-stage, 2nd order method

RungeKutta.TableauRK416 — Function

Tableau of explicit Runge-Kutta method of order four (1/6 rule)

RungeKutta.TableauRK438 — Function

Tableau of explicit Runge-Kutta method of order four (3/8 rule)

RungeKutta.TableauRadauIIA — Function

Radau IIA tableau with s stages

RungeKutta.TableauRalston2 — Function

Tableau of Ralston's two-stage, 2nd order method

RungeKutta.TableauRalston3 — Function

Tableau of Ralston's three-stage, 3rd order method

RungeKutta.TableauRunge — Function

Tableau of Runge's two-stage, 2nd order method

RungeKutta.TableauSRK3 — Function

Tableau of symmetric and symplectic three-stage, 4th order Runge-Kutta method

RungeKutta.TableauSSPRK3 — Function

Tableau of 3rd order Strong Stability Preserving method with three stages

RungeKutta.get_gauss_coefficients — Method

The Gauss coefficients are implicitly given by the so-called simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_gauss_nodes — Method

The Gauss nodes are given by the roots of the shifted Legendre polynomial $P_s (2x-1)$ with $s$ the number of stages.

RungeKutta.get_gauss_weights — Method

The Gauss weights are given by the following integrals

\[b_i = \bigg( \frac{dP}{dx} (c_i) \bigg)^{-2} \int \limits_0^1 \bigg( \frac{P(x)}{x - c_i} \bigg)^2 dx ,\]

where $P(x)$ denotes the shifted Legendre polynomial $P(x) = P_s (2x-1)$ with $s$ the number of stages.

RungeKutta.get_lobatto_coefficients_a — Method

The Lobatto IIIA coefficients are implicitly given by the so-called simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_lobatto_coefficients_b — Method

The Lobatto IIIB coefficients are implicitly given by the so-called simplifying assumption $D(s)$:

\[\sum \limits_{i=1}^{s} b_i c_{i}^{k-1} a_{ij} = \frac{b_j}{k} ( 1 - c_j^k) \qquad j = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_lobatto_coefficients_c — Method

The Lobatto IIIC coefficients are determined by setting $a_{i,1} = b_1$ and solving the so-called simplifying assumption $C(s-1)$, given by

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s-1 ,\]

for $a_{i,j}$ with $i = 1, ..., s$ and $j = 2, ..., s$.

RungeKutta.get_lobatto_coefficients_c̄ — Method

The Lobatto IIIC̄ coefficients are determined by setting $a_{i,s} = 0$ and solving the so-called simplifying assumption $C(s-1)$, given by

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s-1 ,\]

for $a_{i,j}$ with $i = 1, ..., s$ and $j = 1, ..., s-1$.

RungeKutta.get_lobatto_nodes — Method

The s-stage Lobatto nodes are defined as the roots of the following polynomial of degree $s$:

\[\frac{d^{s-2}}{dx^{s-2}} \big( (x - x^2)^{s-1} \big) .\]

RungeKutta.get_lobatto_weights — Method

The Lobatto weights can be explicitly computed by the formula

\[b_j = \frac{1}{s (s-1) P_{s-1}(2 c_j - 1)^2} \qquad j = 1 , \, ... , \, s ,\]

where $P_k$ is the $k$th Legendre polynomial, given by

\[P_k (x) = \frac{1}{k! 2^k} \big( \frac{d^k}{dx^k} (x^2 - 1)^k \big) .\]

RungeKutta.get_radau_coefficients — Method

The Radau IIA coefficients are implicitly given by the so-called simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.get_radau_nodes — Method

The s-stage Radau nodes are defined as the roots of the following polynomial of degree $s$:

\[\frac{d^{s-2}}{dx^{s-2}} \big( (x - x^2)^{s-1} \big) .\]

RungeKutta.get_radau_weights — Method

The Radau weights are implicitly given by the so-called simplifying assumption $B(s)$:

\[\sum \limits_{j=1}^{s} b_{j} c_{j}^{k-1} = \frac{1}{k} \qquad k = 1 , \, ... , \, s .\]

RungeKutta.solve_simplifying_assumption_b — Method

Compute the weights by solving the simplifying assumption $B(s)$:

\[\sum \limits_{j=1}^{s} b_{j} c_{j}^{k-1} = \frac{1}{k} \qquad k = 1 , \, ... , \, s .\]

RungeKutta.solve_simplifying_assumption_c — Method

Compute the coefficients by solving the simplifying assumption $C(s)$:

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{c_i^k}{k} \qquad i = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]

RungeKutta.solve_simplifying_assumption_d — Method

Compute the coefficients by solving the simplifying assumption $D(s)$:

\[\sum \limits_{i=1}^{s} b_i c_{i}^{k-1} a_{ij} = \frac{b_j}{k} ( 1 - c_j^k) \qquad j = 1 , \, ... , \, s , \; k = 1 , \, ... , \, s .\]